Remove old generation module
