+++ pytrackdat/__init__.py
@@ -1,6 +1,12 @@
-from . import common
 from . import analysis
-from . import generation
+from . import common
+from . import design_file
+from . import utils
 
 name = "pytrackdat"
-__all__ = ["common", "analysis", "generation"]
+__all__ = [
+    "analysis",
+    "common",
+    "design_file",
+    "utils",
+]
