+++ pytrackdat/design_file/__init__.py
@@ -1,7 +1,9 @@
 from . import errors
+from . import formatters
 from ._to_relations import design_to_relations
 
 __all__ = [
     "design_to_relations",
     "errors",
+    "formatters",
 ]
