--- pytrackdat/generation/errors.py
@@ -1,26 +1,0 @@
-# PyTrackDat is a utility for assisting in online database creation.
-# Copyright (C) 2018-2021 the PyTrackDat authors.
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#
-# Contact information:
-#     David Lougheed (david.lougheed@gmail.com)
-
-__all__ = [
-    "GenerationError",
-]
-
-
-class GenerationError(Exception):
-    pass
